fix(aero-usb): fix xHCI tests and clippy -D warnings
Recent xHCI refactors and new tests introduced CI failures under the workspace `cargo clippy -- -D warnings` policy.

- Update xHCI MMIO tests to match the current mmio_read/mmio_write APIs.
- Remove clippy-triggering casts/conversions in xHCI tests.
- Avoid `Default` construction for unit structs in tests.

Tests:
- cargo test -p aero-usb --locked
- cargo clippy -p aero-usb --all-targets --all-features -- -D warnings

+++ crates/aero-usb/tests/xhci_enum_smoke.rs
@@ -136,7 +136,7 @@ fn xhci_enum_smoke_bringup_enumerate_and_interrupt_in() {
 
     // --- Capability discovery (driver-style) ---
     let cap0 = xhci.mmio_read(regs::REG_CAPLENGTH_HCIVERSION, 4);
-    let caplength = (cap0 & 0xff) as u64;
+    let caplength = cap0 & 0xff;
     assert_eq!(caplength as u8, regs::CAPLENGTH_BYTES);
 
     let _hcsparams1 = xhci.mmio_read(regs::REG_HCSPARAMS1, 4);
